fin avant herocu

+++ todo/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
+from compte.models import users
 
 ## import todo form and models
 
@@ -14,8 +15,13 @@ def index(request):
 	if request.method == "POST":
 		form = TodoForm(request.POST)
 		if form.is_valid():
-			form.save()
-			return redirect('todo')
+			user = form.cleaned_data['user']
+			if user == request.user:
+				form.save()
+				messages.success(request, "bonne chance a toi ")
+			else: 
+				messages.error(request,"dommage bien esseyé")
+				return redirect('todo')
 	form = TodoForm()
 
 	page = {
@@ -34,7 +40,12 @@ def index(request):
 
 def remove(request, item_id):
 	item = Todo.objects.get(id=item_id)
-	if request.method == "POST":
-		item.delete()
+	is_user = request.user == item.user
+	if is_user or request.user.chef:
+		if request.method == "POST":
+			item.delete()
+			return redirect('todo')
+	else:
+		messages.error(request,"on ne touche pas a se qui n'est pas a soi")
 		return redirect('todo')
 	return render(request,'todo/index.html')
